# docs/contracts/separator-openapi.json
{
  "components": {
    "schemas": {
...
      "CapabilitiesResponse": {
        "properties": {
          "backend": {
            "title": "Backend",
            "type": "string"
          },
          "device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device"
          },
          "features": {
            "$ref": "#/components/schemas/FeatureSet"
          },
          "limits": {
            "$ref": "#/components/schemas/ServerLimits"
          },
          "protocol": {
            "$ref": "#/components/schemas/ProtocolVersion"
          },
          "server_id": {
            "title": "Server Id",
            "type": "string"
          }
        },
        "required": [
          "server_id",
          "protocol",
          "features",
          "backend",
          "device",
          "limits"
        ],
        "title": "CapabilitiesResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "InputRequest": {
        "properties": {
          "filename": {
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "minimum": 1.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "filename",
          "size_bytes",
          "sha256"
        ],
        "title": "InputRequest",
        "type": "object"
      },
      "InputResponse": {
        "properties": {
          "input_id": {
            "title": "Input Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/InputStatus"
          },
          "upload_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Required"
          }
        },
        "required": [
          "input_id",
          "status"
        ],
        "title": "InputResponse",
        "type": "object"
      },
      "InputStatus": {
        "enum": [
          "uploading",
          "ready"
        ],
        "title": "InputStatus",
        "type": "string"
      },
      "InputsResponse": {
        "properties": {
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/InputListItem"
            },
            "title": "Inputs",
            "type": "array"
          }
        },
        "required": [
          "inputs"
        ],
        "title": "InputsResponse",
        "type": "object"
      },
      "JobReferenceResponse": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobReferenceResponse",
        "type": "object"
      },
      "JobRequest": {
        "properties": {
          "force": {
            "default": false,
            "title": "Force",
            "type": "boolean"
          },
          "input_id": {
            "minLength": 1,
            "title": "Input Id",
            "type": "string"
          },
          "model_id": {
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "output_layout": {
            "pattern": "^(two_stem|karaoke)$",
            "title": "Output Layout",
            "type": "string"
          },
          "preset": {
            "pattern": "^(fast|balanced|quality)$",
            "title": "Preset",
            "type": "string"
          }
        },
        "required": [
          "input_id",
          "model_id",
          "preset",
          "output_layout"
        ],
        "title": "JobRequest",
        "type": "object"
      },
...
      "JobsResponse": {
        "properties": {
          "jobs": {
            "items": {
              "discriminator": {
                "mapping": {
                  "cancelled": "#/components/schemas/CancelledJobResponse",
                  "cancelling": "#/components/schemas/RunningJobResponse",
                  "completed": "#/components/schemas/CompletedJobResponse",
                  "failed": "#/components/schemas/FailedJobResponse",
                  "queued": "#/components/schemas/QueuedJobResponse",
                  "running": "#/components/schemas/RunningJobResponse"
                },
                "propertyName": "status"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/QueuedJobResponse"
                },
                {
                  "$ref": "#/components/schemas/RunningJobResponse"
                },
                {
                  "$ref": "#/components/schemas/CompletedJobResponse"
                },
                {
                  "$ref": "#/components/schemas/FailedJobResponse"
                },
                {
                  "$ref": "#/components/schemas/CancelledJobResponse"
                }
              ]
            },
            "title": "Jobs",
            "type": "array"
          }
        },
        "required": [
          "jobs"
        ],
        "title": "JobsResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "K3 Separator Server",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/artifacts/{artifact_id}": {
      "get": {
        "operationId": "download_artifact_v1_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          }
        },
        "summary": "Download Artifact"
      }
    },
    "/v1/capabilities": {
      "get": {
        "operationId": "capabilities_v1_capabilities_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          }
        },
        "summary": "Capabilities"
      }
    },
    "/v1/inputs": {
      "get": {
        "operationId": "list_inputs_v1_inputs_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          }
        },
        "summary": "List Inputs"
      },
      "post": {
        "operationId": "create_input_v1_inputs_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InputRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          }
        },
        "summary": "Create Input"
      }
    },
    "/v1/inputs/{input_id}": {
      "delete": {
        "operationId": "delete_input_v1_inputs__input_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "input_id",
            "required": true,
            "schema": {
              "title": "Input Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          }
        },
        "summary": "Delete Input"
      }
    },
    "/v1/inputs/{input_id}/content": {
      "put": {
        "operationId": "upload_input_v1_inputs__input_id__content_put",
        "parameters": [
          {
            "in": "path",
            "name": "input_id",
            "required": true,
            "schema": {
              "title": "Input Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          }
        },
        "summary": "Upload Input"
      }
    },
    "/v1/jobs": {
      "get": {
        "operationId": "list_jobs_v1_jobs_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          }
        },
        "summary": "List Jobs"
      },
      "post": {
        "operationId": "create_job_v1_jobs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobReferenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          }
        },
        "summary": "Create Job"
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "cancelled": "#/components/schemas/CancelledJobResponse",
                      "cancelling": "#/components/schemas/RunningJobResponse",
                      "completed": "#/components/schemas/CompletedJobResponse",
                      "failed": "#/components/schemas/FailedJobResponse",
                      "queued": "#/components/schemas/QueuedJobResponse",
                      "running": "#/components/schemas/RunningJobResponse"
                    },
                    "propertyName": "status"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/QueuedJobResponse"
                    },
                    {
                      "$ref": "#/components/schemas/RunningJobResponse"
                    },
                    {
                      "$ref": "#/components/schemas/CompletedJobResponse"
                    },
                    {
                      "$ref": "#/components/schemas/FailedJobResponse"
                    },
                    {
                      "$ref": "#/components/schemas/CancelledJobResponse"
                    }
                  ],
                  "title": "Response Get Job V1 Jobs  Job Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          }
        },
        "summary": "Get Job"
      }
    },
    "/v1/jobs/{job_id}/artifacts": {
      "delete": {
        "operationId": "delete_job_artifacts_v1_jobs__job_id__artifacts_delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          }
        },
        "summary": "Delete Job Artifacts"
      }
    },
    "/v1/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobReferenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          }
        },
        "summary": "Cancel Job"
      }
    },
    "/v1/models": {
      "get": {
        "operationId": "models_v1_models_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Machine-readable API error"
          }
        },
        "summary": "Models"
      }
    }
  }
}